+++ core/deferredGraphics/workflows/bloom.cpp
@@ -375,7 +375,7 @@ void bloomGraphics::updateCommandBuffer(uint32_t frameNumber){
     for(uint32_t k = 0; k < frames.size(); k++){
         Texture::transitionLayout(commandBuffers[frameNumber],frames[k].instances[frameNumber].image, VK_IMAGE_LAYOUT_TRANSFER_SRC_OPTIMAL, VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL, VK_REMAINING_MIP_LEVELS, 0, 1);
     }
-    render(commandBuffers[frameNumber], bufferAttachment, frameNumber, frames.size() * image.Count + frameNumber, &bloom);
+    render(commandBuffers[frameNumber], bufferAttachment, frameNumber, static_cast<uint32_t>(frames.size()) * image.Count + frameNumber, &bloom);
     Texture::transitionLayout(commandBuffers[frameNumber], blitBufferImage, VK_IMAGE_LAYOUT_TRANSFER_SRC_OPTIMAL, VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL, VK_REMAINING_MIP_LEVELS, 0, 1);
 }
